[ADD] internal: Introduce commit selection delegation logic
The file `internal/tui/components.go` was modified to introduce a function to
delegate responsibility in the type of commit selection.

The file `go.mod` was updated to include a new dependency:
`github.com/dustin/go-humanize v1.0.1`. The file `go.sum` was also updated to
reflect this change. The file `internal/tui/model.go` was updated to make use of
the new dependency. The  file `internal/tui/update.go` was updated to introduce
functions for handling the logic related to creating a commit. Lastly, the file
`internal/tui/view.go` was updated to modify the view depending on the state of
building a commit.