[IMP] tui: Conditional styling for TUI files
The file `internal/tui/components.go` was modified to add
conditional styling based on the type of file. A new check
was introduced to determine if a file is executable. Based on
the file type, the `baseStyle` is set to a corresponding color.

There is no change to other files.