[ADD] utils.go: Add new GetStagedDiffSummary function in TUI utils
The internal/tui/utils.go file was modified to add a new function
GetStagedDiffSummary, which generates a string containing the diffs of all
staged files in a git repository.

This function uses git commands to get the list of staged files and their diffs.

The file internal/tui/utils.go was also modified to add new imports, including
os/exec.