[REF] tui: Change commit view layout
The file `internal/tui/view.go` was modified to change the layout of the commit view. The `model.WritingStatusBar.Render()` element was moved to be the first element in the vertical layout, and its rendering was made to occupy the top space. The styles for `VerticalSpace` were introduced, defining a height of 1 unit.

The file `internal/tui/view.go` was also modified to remove the rendering of `selectedCommitTypeStyle` and related variables. The `uiElements` variable now replaces `mainContent` in the vertical layout, and `VerticalSpace` is used to insert empty lines between elements.