[IMP] tui: Add status bar update functionality
The file `internal/tui/update.go` was modified to update the `WritingStatusBar` model with new content. Specifically, the `updateChoosingScope` function was updated to set `model.WritingStatusBar.Level` to `statusbar.LevelInfo` and `model.WritingStatusBar.Content` to a formatted string with the path of the chosen file or folder. Additionally, an error message was updated to be displayed in the status bar with `model.WritingStatusBar.Level` set to `statusbar.LevelError` and `model.WritingStatusBar.Content` set to a plain text error message.

The `updateChoosingScope` function now displays file system information on the status bar, whereas previously a specific error message could be displayed by setting `model.WritingStatusBar.Level` to `statusbar.LevelError` and `model.WritingStatusBar.Content` set to a specific error message. The `github.com/charmbracelet/lipgloss/v2` package was commented out, and `model.fileList.NewStatusMessage` and a line setting `model.WritingStatusBar.Content` to a plain text message were removed.