[FIX] tui: Adjust main list height
The file `internal/tui/view.go` was modified to adjust the height of the main list. Specifically, calculations were added to determine the available height for the list, taking into account the heights of the status bar and vertical space. The `model.mainList.SetSize` function was called with the updated available height.

No other changes were found in this modification.