[FIX] tui: Fix `tui` visibility in history list
The file `internal/tui/components.go` was modified to make the filter visible in the main list. Specifically, the `SetShowFilter` method of `historyList` was updated to display the filter by setting its argument to `true`.

The `FilterValue` method of `HistoryCommitItem` was updated to remove the inclusion of the date string and commit ID as part of the filter value, previously formatted as `hci.commit.CreatedAt.Format("2006-01-02") + " " + fmt.Sprintf("%d", hci.commit.ID) + " ")`.