[FIX] tui: Record commit creation message in status bar
The file `internal/tui/update.go` was modified to change the behavior of the commit creation message. Specifically, the `createCommit` function now uses `model.WritingStatusBar.ShowMessageForDuration` to display the "Record created in the db successfully" message for a duration of 2 seconds, instead of leaving it indefinitely. The message is also replaced with a new string that includes the main list title.