[ADD] storage: Database release support
The files `internal/storage/database.go`, `internal/storage/queries.go`, and `internal/storage/types.go` were modified to support the addition of a new table for releases. In `database.go`, a new table `releases` was created with columns for `id`, `title`, `body`, `branch`, `commit_list`, `version`, `workspace`, and `created_at`. In `queries.go`, two new functions were added: `CreateRelease` to insert a new release into the `releases` table, and `GetReleases` to retrieve a list of releases for a given workspace. In `types.go`, a new struct `Release` was defined to represent a release.